requirements.txt
watchdog
pdf2image
Pillow
pywin32
customtkinter
requests
pyinstaller
